organize unit tests
